Move `Opaque` to `stable_mir`
